src/requirements.txt
fastapi
pydantic
motor
uvicorn
python-dotenv
email-validator